Changes in BT Permissions.
The way local file access works is it prompts user, and it will use that response forever until changed in app settings

For Bluetooth, it is the same now— Single prompt once app is opened; based .

However, we might incorporate a settings button. In that settings button, we’re able to toggle on/off file access/bluetooth access for ease of use.